+++ src/components/OrderPage/OrderForm.tsx
@@ -1,8 +1,7 @@
 import styled from '@emotion/styled'
 import { cardMock } from '@/pages/OrderPage/cardMock'
 import type { Product } from '@/types/product'
-import { useState } from 'react'
-import { useForm, useFieldArray } from 'react-hook-form'
+import { useForm } from 'react-hook-form'
 
 import { CardSelector } from './CardSelector'
 import { SelectedCard } from './SelectedCard'
@@ -10,6 +9,8 @@ import { SenderInput } from './SenderInput'
 import { ReceiverSection } from './ReceiverSection'
 import { ReceiverModal } from './ReceiverModal'
 import { ProductInfo } from './ProductInfo'
+
+import { useOrderForm } from '@/hooks/useOrderForm'
 
 interface OrderFormProps {
   product: Product
@@ -27,16 +28,7 @@ export interface FormValues {
 }
 
 export function OrderForm({ product }: OrderFormProps) {
-  const {
-    register,
-    handleSubmit,
-    setValue,
-    control,
-    reset,
-    getValues,
-    trigger,
-    formState: { errors },
-  } = useForm<FormValues>({
+  const form = useForm<FormValues>({
     mode: 'onChange',
     defaultValues: {
       sender: '',
@@ -45,50 +37,26 @@ export function OrderForm({ product }: OrderFormProps) {
     },
   })
 
-  const { fields, append, remove } = useFieldArray({
-    control,
-    name: 'receivers',
-  })
+  const {
+    fields,
+    append,
+    remove,
+    showReceiverModal,
+    setShowReceiverModal,
+    selectedCard,
+    setSelectedCard,
+    finalReceivers,
+    totalPrice,
+    onSubmit,
+    validateAndSaveReceivers,
+  } = useOrderForm(form, product)
 
-  const [showReceiverModal, setShowReceiverModal] = useState(false)
-  const [selectedCard, setSelectedCard] = useState(cardMock[0])
-  const [finalReceivers, setFinalReceivers] = useState<ReceiverInfo[]>([])
-
-  const totalQuantity = finalReceivers.reduce(
-    (acc, receiver) => acc + Number(receiver.quantity || 0),
-    0
-  )
-  const totalPrice = product.price.sellingPrice * totalQuantity
-
-  const onSubmit = (data: FormValues) => {
-    alert(`주문이 완료되었습니다.
-      상품명: ${product.name}
-      구매 수량: ${totalQuantity}
-      발신자 이름: ${data.sender}
-      메시지: ${data.message}`)
-    reset()
-    setSelectedCard(cardMock[0])
-    setValue('message', cardMock[0].defaultTextMessage)
-    setFinalReceivers([])
-  }
-
-  const validateAndSaveReceivers = async () => {
-    const isValid = await trigger('receivers')
-
-    if (!isValid) return
-
-    const receivers = getValues('receivers')
-    const phones = receivers.map((r) => r.phone)
-    const hasDuplicatePhone = new Set(phones).size !== phones.length
-
-    if (hasDuplicatePhone) {
-      alert('중복된 전화번호가 있습니다.')
-      return
-    }
-
-    setFinalReceivers(receivers)
-    setShowReceiverModal(false)
-  }
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+    setValue,
+  } = form
 
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
